Internals: Replace unsigned short with ImWchar when dealing with character storage (missing cases) + in imgui_impl_ file keep using neutral type everywhere, added missing explicit cast in three bindings. (#2078)

+++ examples/imgui_impl_osx.mm
@@ -188,7 +188,7 @@ bool ImGui_ImplOSX_HandleEvent(NSEvent* event, NSView* view)
         {
             int c = [str characterAtIndex:i];
             if (c < 0xF700 && !io.KeyCtrl)
-                io.AddInputCharacter(c);
+                io.AddInputCharacter((unsigned short)c);
 
             // We must reset in case we're pressing a sequence of special keys while keeping the command pressed
             int key = mapCharacterToKey(c);
